requirements.txt
Django
djangorestframework
mysqlclient
python-dotenv
openai
django-cors-headers
